Fixes images routes
